// src/app/interface/design/design-view.interface.ts
export interface IDesignView {
  _id?: string;
  designName?: string;
  colorOne?: string;
  colorTwo?: string;
  colorThree?: string;
  colorFour?: string;  
  _profilePic?: string;
  _profilePicUrl?: string;
  _profilePicSecureUrl?: string;
  _host?: string;
  _hostName?: string;
  _hostEmail?: string;
}